+++ packages/cqrs-bundle/CqrsBundle.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SonsOfPHP\Bundle\Cqrs;
+
+use Symfony\Component\HttpKernel\Bundle\AbstractBundle;
+
+/**
+ * @experimental
+ */
+class CqrsBundle extends AbstractBundle
+{
+}